[DC-2003] Remove functions from constants files and consolidate
